+++ app/src/main/java/com/example/afya/viewmodel/PostViewModel.kt
@@ -1,0 +1,53 @@
+package com.example.afya.viewmodel
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.example.afya.model.Post
+import com.example.afya.repository.PostRepository
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.launch
+
+data class UIState(
+    val isLoading: Boolean = false,
+    val posts: List<Post> = emptyList(),
+    val error: String? = null
+)
+
+class PostViewModel : ViewModel() {
+   
+ 
+    private val _uiState = MutableStateFlow(UIState())
+    val uiState: StateFlow<UIState> = _uiState.asStateFlow()
+
+    init{
+        loadPosts()
+    }
+
+
+    fun loadPosts() {
+     
+        _uiState.value = _uiState.value.copy(isLoading = true)
+        viewModelScope.launch {
+            try {
+                delay(2000)
+                val posts = PostRepository.getPosts()
+                _uiState.value = _uiState.value.copy(
+                    isLoading = false,
+                    posts = posts
+                )
+            } catch (e: Exception) {
+                _uiState.value = _uiState.value.copy(
+                    isLoading = false,
+                    error = e.message
+                )
+            }
+        }
+    }
+        
+}
+
